constantes, metodos e atributos estaticos, tratamento de excecoes

+++ Java 2TDSG/01-Revisao/src/br/com/fiap/bean/IContaInvestimento.java
@@ -2,5 +2,5 @@ package br.com.fiap.bean;
 
 public interface IContaInvestimento {
 
-	float calculaRetornoInvestimento();
+	double calculaRetornoInvestimento();
 }
